# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Civic Cleanup Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports:
    get:
      operationId: getReports
      tags: [reports]
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: citizenId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Reports list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportsResponse"
    post:
      operationId: createReport
      tags: [reports]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReportBody"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResponse"
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResponse"
    patch:
      operationId: updateReport
      tags: [reports]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReportBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResponse"
...
  /reports/verify:
    post:
      operationId: verifyReport
      tags: [reports]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyReportBody"
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  # WALLET
...
  /scrap/listings:
    get:
      operationId: getScrapListings
      tags: [scrap]
      parameters:
        - name: citizenId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Scrap listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrapListingsResponse"
    post:
      operationId: createScrapListing
      tags: [scrap]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScrapListingBody"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrapListingResponse"
  /scrap/listings/{id}:
    get:
      operationId: getScrapListing
      tags: [scrap]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrapListingResponse"
    patch:
      operationId: updateScrapListing
      tags: [scrap]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScrapListingBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrapListingResponse"

  # DONATIONS
...
  /dustbin/collections:
    get:
      operationId: getDustbinCollections
      tags: [reports]
      responses:
        "200":
          description: Collections
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DustbinCollectionsResponse"
    post:
      operationId: createDustbinCollection
      tags: [reports]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDustbinCollectionBody"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  # DASHBOARD SUMMARY
...
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: ["string", "null"]
        pointsEarned:
          type: ["integer", "null"]
      required: [ok]

    # AUTH
...
    WasteReport:
      type: object
      properties:
        id:
          type: string
        citizenId:
          type: string
        imageUrl:
          type: ["string", "null"]
        wasteType:
          type: string
        description:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        address:
          type: ["string", "null"]
        status:
          type: string
        assignedWorkerId:
          type: ["string", "null"]
        completionImageUrl:
          type: ["string", "null"]
        rewardPoints:
          type: ["integer", "null"]
        priority:
          type: ["string", "null"]
        reporterName:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, citizenId, wasteType, status, createdAt]
    ReportResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/WasteReport"
      required: [data]
    ReportsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/WasteReport"
      required: [data]
    CreateReportBody:
      type: object
      properties:
        imageUrl:
          type: string
        wasteType:
          type: string
        description:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        address:
          type: string
        priority:
          type: string
      required: [wasteType]
    UpdateReportBody:
      type: object
      properties:
        status:
          type: string
        assignedWorkerId:
          type: string
        completionImageUrl:
          type: string
        priority:
          type: string
    VerifyReportBody:
      type: object
      properties:
        reportId:
          type: string
        isLegit:
          type: boolean
      required: [reportId, isLegit]

    # WALLET
...
    ScrapListingItem:
      type: object
      properties:
        id:
          type: string
        listingId:
          type: string
        itemName:
          type: string
        category:
          type: string
        weightKg:
          type: number
        pricePerKg:
          type: number
        totalPrice:
          type: ["number", "null"]
      required: [id, listingId, itemName, category, weightKg, pricePerKg]
    ScrapListing:
      type: object
      properties:
        id:
          type: string
        citizenId:
          type: string
        status:
          type: string
        dealerId:
          type: ["string", "null"]
        totalEstimate:
          type: ["number", "null"]
        totalWeight:
          type: ["number", "null"]
        address:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/ScrapListingItem"
        createdAt:
          type: string
      required: [id, citizenId, status, createdAt]
    ScrapListingResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/ScrapListing"
      required: [data]
    ScrapListingsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ScrapListing"
      required: [data]
    CreateScrapListingBody:
      type: object
      properties:
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        items:
          type: array
          items:
            type: object
            properties:
              itemName:
                type: string
              category:
                type: string
              weightKg:
                type: number
              pricePerKg:
                type: number
            required: [itemName, category, weightKg, pricePerKg]
      required: [items]
    UpdateScrapListingBody:
      type: object
      properties:
        status:
          type: string
        dealerId:
          type: string

    # DONATIONS
...
    DustbinCollection:
      type: object
      properties:
        id:
          type: string
        citizenId:
          type: string
        workerId:
          type: ["string", "null"]
        fillLevel:
          type: ["integer", "null"]
        pointsAwarded:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, citizenId, createdAt]
    DustbinCollectionsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/DustbinCollection"
      required: [data]
    CreateDustbinCollectionBody:
      type: object
      properties:
        citizenId:
          type: string
        fillLevel:
          type: integer
        notes:
          type: string
      required: [citizenId, fillLevel]

    # DASHBOARD SUMMARIES
...